Misc: lots of tidying.
