+++ src/ship_space.cc
@@ -108,9 +108,9 @@ ship_space::get_chunk_containing(int block_x, int block_y, int block_z)
 {
     int chunk_x, chunk_y, chunk_z;
 
-    split_coord(block_x, NULL, &chunk_x);
-    split_coord(block_y, NULL, &chunk_y);
-    split_coord(block_z, NULL, &chunk_z);
+    split_coord(block_x, nullptr, &chunk_x);
+    split_coord(block_y, nullptr, &chunk_y);
+    split_coord(block_z, nullptr, &chunk_z);
 
     return this->get_chunk(chunk_x, chunk_y, chunk_z);
 }
@@ -128,7 +128,7 @@ ship_space::get_chunk(int chunk_x, int chunk_y, int chunk_z)
         return it->second;
     }
 
-    return NULL;
+    return nullptr;
 }
 
 
@@ -173,7 +173,7 @@ ship_space::raycast(glm::vec3 o, glm::vec3 d, raycast_info *rc)
     int ny = 0;
     int nz = 0;
 
-    block *bl = 0;
+    block *bl = nullptr;
 
     bl = this->get_block(x,y,z);
     rc->inside = bl ? bl->type != block_empty : 0;
